+++ post-processor/exoscale-import/post-processor_acc_test.go
@@ -20,6 +20,9 @@ var (
 	testAccTemplateName        = "packer-plugin-test-" + new(testSuite).randomString(6)
 	testAccTemplateDescription = new(testSuite).randomString(10)
 	testAccTemplateUsername    = strings.ToLower(new(testSuite).randomString(10))
+	testAccTemplateMaintainer  = "Exoscale"
+	testAccTemplateVersion     = "0.acceptance"
+	testAccTemplateBuild       = new(testSuite).randomString(8)
 	testAccImageFile           = "./testdata/test-packer-post-processor-exoscale-import.qcow2"
 )
 
@@ -55,6 +58,9 @@ func TestAccPostProcessor(t *testing.T) {
 		"template_name":        testAccTemplateName,
 		"template_description": testAccTemplateDescription,
 		"template_username":    testAccTemplateUsername,
+		"template_maintainer":  testAccTemplateMaintainer,
+		"template_version":     testAccTemplateVersion,
+		"template_build":       testAccTemplateBuild,
 	}}...)
 	require.NoError(t, err)
 
@@ -73,6 +79,9 @@ func TestAccPostProcessor(t *testing.T) {
 		require.Equal(t, testAccTemplateDescription, *template.Description)
 		require.Equal(t, defaultTemplateBootMode, *template.BootMode)
 		require.Equal(t, testAccTemplateUsername, *template.DefaultUser)
+		require.Equal(t, testAccTemplateMaintainer, *template.Maintainer)
+		require.Equal(t, testAccTemplateVersion, *template.Version)
+		require.Equal(t, testAccTemplateBuild, *template.Build)
 	}
 
 	require.NoError(t, artifact.Destroy())