+++ post-processor/exoscale-import/step_register_template_test.go
@@ -17,6 +17,9 @@ func (ts *testSuite) TestStepRegisterTemplate_Run() {
 			TemplateDescription: ts.randomString(10),
 			TemplateUsername:    ts.randomString(10),
 			TemplateBootMode:    ts.randomString(10),
+			TemplateMaintainer:  ts.randomString(10),
+			TemplateVersion:     ts.randomString(10),
+			TemplateBuild:       ts.randomString(10),
 		}
 		testTemplatePasswordEnabled = !testConfig.TemplateDisablePassword
 		testTemplateSSHKeyEnabled   = !testConfig.TemplateDisableSSHKey
@@ -27,13 +30,16 @@ func (ts *testSuite) TestStepRegisterTemplate_Run() {
 
 	testTemplate := &egoscale.Template{
 		BootMode:        &testConfig.TemplateBootMode,
+		Build:           &testConfig.TemplateBuild,
 		Checksum:        &testSnapshotChecksum,
 		DefaultUser:     &testConfig.TemplateUsername,
 		Description:     &testConfig.TemplateDescription,
+		Maintainer:      &testConfig.TemplateMaintainer,
 		Name:            &testConfig.TemplateName,
 		PasswordEnabled: &testTemplatePasswordEnabled,
 		SSHKeyEnabled:   &testTemplateSSHKeyEnabled,
 		URL:             &testSnapshotPresignedURL,
+		Version:         &testConfig.TemplateVersion,
 	}
 	ts.state.Put("image_checksum", testSnapshotChecksum)
 	ts.state.Put("image_url", testSnapshotPresignedURL)
@@ -72,6 +78,9 @@ func (ts *testSuite) TestStepRegisterTemplateWithEmptyFields_Run() {
 			TemplateDescription: "",
 			TemplateUsername:    "",
 			TemplateBootMode:    ts.randomString(10),
+			TemplateMaintainer:  "",
+			TemplateVersion:     "",
+			TemplateBuild:       "",
 		}
 		testTemplatePasswordEnabled = !testConfig.TemplateDisablePassword
 		testTemplateSSHKeyEnabled   = !testConfig.TemplateDisableSSHKey
@@ -82,13 +91,16 @@ func (ts *testSuite) TestStepRegisterTemplateWithEmptyFields_Run() {
 
 	testTemplate := &egoscale.Template{
 		BootMode:        &testConfig.TemplateBootMode,
+		Build:           nil,
 		Checksum:        &testSnapshotChecksum,
 		DefaultUser:     nil,
 		Description:     nil,
+		Maintainer:      nil,
 		Name:            &testConfig.TemplateName,
 		PasswordEnabled: &testTemplatePasswordEnabled,
 		SSHKeyEnabled:   &testTemplateSSHKeyEnabled,
 		URL:             &testSnapshotPresignedURL,
+		Version:         nil,
 	}
 	ts.state.Put("image_checksum", testSnapshotChecksum)
 	ts.state.Put("image_url", testSnapshotPresignedURL)